Fix issue with some image not showing if there's only one:
One more issue (sorry haha), on most, but somehow not all pages, the project 3 tab needs you to click next in order for the single picture to show up. Here's an example: https://makeitbig.calarts.edu/html/student/bfa3/joe-suh.html
Joe's Children of Men poster doesn't show up, but it's there and none of the code is any different.

I have been putting two of the same images in some of these pages because then it loads, but if there is a proper fix let me know.